+++ backend/app/rest/api/ssl.go
@@ -1,7 +1,6 @@
 package api
 
 import (
-	"context"
 	"crypto/tls"
 	"fmt"
 	"net/http"
@@ -63,26 +62,6 @@ func (s *Rest) httpChallengeRouter(m *autocert.Manager) http.Handler {
 
 	router.Handle("/", m.HTTPHandler(s.redirectHandler()))
 	return router
-}
-
-// timeout returns a middleware matching chi's middleware.Timeout: it sets a
-// deadline on the request context and writes 504 Gateway Timeout if the
-// deadline is exceeded. The 504 is sent once the downstream handler returns
-// after observing the canceled context; a handler that ignores r.Context()
-// is not aborted.
-func timeout(d time.Duration) func(http.Handler) http.Handler {
-	return func(next http.Handler) http.Handler {
-		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			ctx, cancel := context.WithTimeout(r.Context(), d)
-			defer func() {
-				cancel()
-				if ctx.Err() == context.DeadlineExceeded {
-					w.WriteHeader(http.StatusGatewayTimeout)
-				}
-			}()
-			next.ServeHTTP(w, r.WithContext(ctx))
-		})
-	}
 }
 
 func (s *Rest) redirectHandler() http.Handler {
